kafka-webview-ui/src/main/java/org/sourcelab/kafka/webview/ui/manager/kafka/deserializer/BytesToHexDeserializer.java
/**
 * MIT License
 *
 * Copyright (c) 2017-2022 SourceLab.org (https://github.com/SourceLabOrg/kafka-webview/)
 *
 * Permission is hereby granted, free of charge, to any person obtaining a copy
 * of this software and associated documentation files (the "Software"), to deal
 * in the Software without restriction, including without limitation the rights
 * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
 * copies of the Software, and to permit persons to whom the Software is
 * furnished to do so, subject to the following conditions:
 *
 * The above copyright notice and this permission notice shall be included in all
 * copies or substantial portions of the Software.
 *
 * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
 * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
 * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
 * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
 * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 * SOFTWARE.
 */

package org.sourcelab.kafka.webview.ui.manager.kafka.deserializer;

import org.apache.kafka.common.serialization.Deserializer;
import org.apache.tomcat.util.buf.HexUtils;

import java.util.Map;

/**
 * Deserializes bytes into Hex strings.  Intended to allow for a more user friendly display of bytes.
 */
public class BytesToHexDeserializer implements Deserializer<String> {

    @Override
    public void configure(final Map<String, ?> configs, final boolean isKey) {
    }

    @Override
    public String deserialize(final String topic, final byte[] data) {
        // Convert bytes into Hex.
        return HexUtils.toHexString(data);
    }

    @Override
    public void close() {
    }
}
